+++ CHANGELOG.md
@@ -11,12 +11,13 @@ and this project adheres to
 - 🚸(frontend) allow opening "@page" links with ctrl/command/middle-mouse click
 - ✅ E2E - Any instance friendly #2142
 
+### Changed
+
+- ♻️(backend) do not paginate threads list response #2186
+
 ### Fixed
 
 - 🐛(frontend) Fix drop cursor creating columns #2185
-
-### Fixed
-
 - 🐛 Fixed side effects between comments and versioning #2183
 
 ## [v4.8.5] - 2026-04-03